requirements (1).txt
scikit-learn==1.1.1
Pillow==9.1.0
pyTelegramBotAPI==4.6.1
tensorflow==2.6.0
numpy==1.22.3